# api-documentation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/project/requirement/{requirementId}/comment": {
      "post": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "createComment",
        "parameters": [
          {
            "name": "requirementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/project/requirement/{requirementId}/comment/{commentId}": {
      "post": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "rateComment",
        "parameters": [
          {
            "name": "requirementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRatingRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentRatingResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentRequestDTO": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        }
      },
      "CommentResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "requirementId": {
            "type": "integer",
            "format": "int64"
          },
          "creationDate": {
            "type": "string",
            "format": "date"
          },
          "likeCount": {
            "type": "integer",
            "format": "int64"
          },
          "dislikeCount": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string"
          },
          "userHasRated": {
            "type": "string",
            "enum": [
              "LIKE",
              "DISLIKE"
            ]
          }
        }
      },
      "CommentRatingRequestDTO": {
        "type": "object",
        "properties": {
          "commentRatingType": {
            "type": "string",
            "enum": [
              "LIKE",
              "DISLIKE"
            ]
          }
        }
      },
      "CommentRatingResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "commentId": {
            "type": "integer",
            "format": "int64"
          },
          "commentRatingType": {
            "type": "string",
            "enum": [
              "LIKE",
              "DISLIKE"
            ]
          }
        }
      },
...
    }
  }
}